+++ R/variables.R
@@ -2,3 +2,17 @@
 # change with each hunting season
 
 REF_CURRENT_SEASON <- "2024"
+
+REF_RELEASES <-
+  c("incomplete" = "1.2.0",
+    "incomplete" = "1.2.1",
+    "incomplete" = "1.2.2",
+    "incomplete" = "1.2.3",
+    `2021-2022` = "1.2.4",
+    "incomplete" = "1.2.5",
+    "incomplete" = "1.2.6",
+    `2022-2023` = "1.2.7",
+    `2023-2024` = "1.2.8",
+    `2024-2025` = "1.3.0",
+    `2025-2026` = "1.4.0"
+  )